refactor(tests): split Playwright example tests by tab
Split the monolithic app.spec.ts (1182 lines, 88 tests) into 11 focused
spec files plus a shared helpers module:

- helpers.ts: goToApp, clickTab, tabIds
- app.spec.ts: app shell (heading, tabs)
- counter.spec.ts, todos.spec.ts, theme.spec.ts, data-fetcher.spec.ts,
  hooks-showcase.spec.ts, shown.spec.ts, ui-patch.spec.ts (incl. both
  visibility sections), lazy-context.spec.ts, abort-signal.spec.ts,
  context-scoping.spec.ts

Total test count: 88 (unchanged). All 264 cross-browser runs pass.

Closes #64
